backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
httpx>=0.27.0
python-dotenv>=1.0.0
cryptography>=43.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
mysql-connector-python>=8.0.33

# Optional STT/TTS
# openai-whisper==20231117
# vosk==0.3.45
# pyttsx3==2.90
# TTS==0.22.0

# Optional wake word
# pvporcupine==3.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
